==> code/01b_direct_estimates_nofilter.Rmd <==
---
title: "Direct Estimates"
---

Summary: Compute direct estimates of standard errors using the survey_mean function with beta option. 

```{r}
## load packages and helper functions 
library(here)
source(here("code", "helpers.R"))
```

```{r}
## read in general survey
gensample <- read_csv("../data/full_survey.csv", guess_max = 15000, show_col_types = FALSE) %>% 
  filter(p_over == 1)

## oversample
oversample <- read_csv("../data/full_survey.csv", guess_max = 15000, show_col_types = FALSE)

recode_nofilter <- function(df) {
  
  df <- df %>% 
    mutate(
      banking = case_when(
        ! is.na(banking_flag) ~ 1, 
        TRUE ~ 0),
      banking_lostmoney = case_when(
        banking_money > 0 ~ 1, 
        TRUE ~ 0),
        banking = case_when(
        ! is.na(banking_flag) ~ 1, 
        TRUE ~ 0),
      nondelivery = case_when(
        ! is.na(nondelivery_flag) ~ 1, 
        TRUE ~ 0),
      nonpayment = case_when(
        ! is.na(nonpayment_flag) ~ 1, 
        TRUE ~ 0),
      overpayment = case_when(
        ! is.na(overpayment_flag) ~ 1, 
        TRUE ~ 0),
      advanced_fee = case_when(
        ! is.na(advanced_fee_flag) ~ 1, 
        TRUE ~ 0),
      extortion = case_when(
        ! is.na(extortion_flag) ~ 1, 
        TRUE ~ 0),
    )
  return(df)
}

gensample <- recode_nofilter(df = gensample)
oversample <- recode_nofilter(df = oversample)
```


## Full Sample 

```{r}
## create survey object 
dsurvey <- gensample %>% 
  as_survey_design(weight = weight)

## calculate survey means using proportion argument 
direct_estimates <- dsurvey %>% 
  summarize(banking.lostmoney = survey_mean(banking_lostmoney, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            banking = survey_mean(banking, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            nondelivery = survey_mean(nondelivery, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            nonpayment = survey_mean(nonpayment, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            overpayment = survey_mean(overpayment, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            extortion = survey_mean(extortion, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            advanced.fee = survey_mean(advanced_fee, proportion = T, prop_method = "beta", vartype = c("ci", "se"))) %>% 
  mutate(id = "id") %>% 
  pivot_longer(- id) %>% 
  separate(name, c("scam", "type"), sep = "_", fill = "right") %>% 
  pivot_wider(names_from = c(type), values_from = value, names_sep = "_") %>% 
  select(scam,
         estimate = `NA`,
         se = se,
         low,
         upp) %>% 
  arrange(desc(estimate)) %>% 
  mutate(scam = case_when(
    scam == "banking" ~ "Banking / Credit Card",
    scam == "nondelivery" ~ "Non-Delivery",
    scam == "banking.lostmoney" ~ "Banking / Credit Card (Lost Money)",
    scam == "advanced.fee" ~ "Advanced Fee",
    scam == "nonpayment" ~ "Nonpayment",
    scam == "extortion" ~ "Extortion",
    scam == "overpayment" ~ "Overpayment"
  ))

## create survey object 
oversample_survey <- oversample %>% 
  as_survey_design(weight = weight2)

## calculate survey means using proportion argument 
oversample_summary_stats <- oversample_survey %>% 
  summarize(
            nonpayment = survey_mean(nonpayment, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            overpayment = survey_mean(overpayment, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            extortion = survey_mean(extortion, proportion = T, prop_method = "beta", vartype = c("ci", "se")),
            advanced.fee = survey_mean(advanced_fee, proportion = T, prop_method = "beta", vartype = c("ci", "se"))
  ) %>% 
  mutate(id = "id") %>% 
  pivot_longer(- id) %>% 
  separate(name, c("scam", "type"), sep = "_", fill = "right") %>% 
  pivot_wider(names_from = c(type), values_from = value, names_sep = "_") %>% 
  select(scam,
         mean = `NA`,
         se = se,
         low,
         upp) %>% 
  arrange(desc(mean)) %>% 
  mutate()

## final direct results for the paper  
## Oversample for advanced fee, extortion, nonpayment, and overpayment 
direct_estimates <- oversample_summary_stats %>% 
  rename(estimate = mean) %>% 
  bind_rows(direct_estimates %>% 
              filter(scam %in% c("Banking / Credit Card", "Non-Delivery", "Banking / Credit Card (Lost Money)"))) %>% 
  mutate(scam = case_when(
    scam == "advanced.fee" ~ "Advanced Fee",
    scam == "nonpayment" ~ "Non-Payment",
    scam == "extortion" ~ "Extortion",
    scam == "overpayment" ~ "Overpayment", 
    TRUE ~ scam
  )) %>% 
  mutate(estimate_annualized = 1 - (1-estimate)^(1/2),
         low_annualized = 1 - (1-low)^(1/2),
         upper_annualized = 1 - (1-upp)^(1/2),
         method = "Direct")
```

```{r}
write_csv(direct_estimates, file = "../data/results/direct_estimates_nofilter.csv")
```
