==> code/03b_scaleup_estimates.Rmd <==
---
title: "Network Scale-up Estimates"
---

Network Scale-up Estimates — No Response Verification 

```{r}
## load packages and helper functions 
library(here)
source(here("code", "helpers.R"))

## set seed for reproducibility
set.seed(1)
```

## Direct Network Scale-up Estimates 
```{r}
## read in survey
survey <- read_csv("../data/full_survey.csv", guess_max = 15000, show_col_types = FALSE) 

recode_nofilter <- function(df) {
  
  df <- df %>% 
    mutate(
      banking = case_when(
        ! is.na(banking_flag) ~ 1, 
        TRUE ~ 0),
      banking_lostmoney = case_when(
        banking_money > 0 ~ 1, 
        TRUE ~ 0),
        banking = case_when(
        ! is.na(banking_flag) ~ 1, 
        TRUE ~ 0),
      nondelivery = case_when(
        ! is.na(nondelivery_flag) ~ 1, 
        TRUE ~ 0),
      nonpayment = case_when(
        ! is.na(nonpayment_flag) ~ 1, 
        TRUE ~ 0),
      overpayment = case_when(
        ! is.na(overpayment_flag) ~ 1, 
        TRUE ~ 0),
      advanced_fee = case_when(
        ! is.na(advanced_fee_flag) ~ 1, 
        TRUE ~ 0),
      extortion = case_when(
        ! is.na(extortion_flag) ~ 1, 
        TRUE ~ 0),
    )
  return(df)
}

survey <- recode_nofilter(survey)

## read in populations of known size 
known.size <- read_csv("../data/known_popsize/groups_known_size.csv",  show_col_types = FALSE) %>% 
  mutate(name = tolower(paste0("name_", name)))
```

```{r}
## top code number at 30
survey <- survey %>%
  mutate(across(c(starts_with("name")), top_code))

## Network Size
survey <- survey %>% 
  mutate(network_size = name_emily + name_rachel + name_paula + name_martha + name_tina + name_rose +
           name_adam + name_alan + name_walter + name_bruce + name_kyle + name_ralph)

## estimate of adults in the US over age 18 
tot.pop.size <- 328239523 * (1-0.223)

## set vector so we can use with our package
kp.vec <- df.to.kpvec(known.size, kp.var="name", kp.value="size")

## estimate person network size
degree <- kp.degree.estimator(survey.data=survey,
                              known.popns=kp.vec,
                              total.popn.size=tot.pop.size,
                              missing="complete.obs")

survey$degree <- degree

survey_subsample <- survey %>% 
  filter(p_over == 1) 

## manually compute scale-up estimates (c)
basic_nsum_results <- survey_subsample %>% 
  summarize(banking = sum(scaleup_banking*weight)/sum(degree*weight),
            banking_lostmoney = sum(scaleup_banking_desc_lost_money_count*weight, na.rm = T)/sum(degree*weight),
            nondelivery = sum(scaleup_nondelivery*weight)/sum(degree*weight),
            nonpayment = sum(scaleup_nonpayment*weight)/sum(degree*weight),
            overpayment = sum(scaleup_overpayment*weight)/sum(degree*weight),
            advanced_fee = sum(scaleup_advanced_fee*weight)/sum(degree*weight),
            extortion = sum(scaleup_extortion*weight)/sum(degree*weight)) %>% 
  mutate(id = "id") %>% 
  pivot_longer(-id) %>% 
  select(scam = name, proportion = value)

```


  scam              proportion
  <chr>                  <dbl>
1 banking           0.000677  
2 banking_lostmoney 0.000139  
3 nondelivery       0.000102  
4 nonpayment        0.00000623
5 overpayment       0.0000118 
6 advanced_fee      0.00000657
7 extortion         0.0000219

## Generalized Scale-up

```{r}
###############################################
#### Rare Individual Adjustment Factors #######
###############################################

## Individual Degree Ratio
degree_ratio_oversample <-  survey %>% 
  summarize(nonpayment = mean(degree[nonpayment == 1] * weight2[nonpayment == 1], na.rm = T) / mean(degree * weight2, na.rm = T),
            overpayment = mean(degree[overpayment == 1] * weight2[overpayment == 1], na.rm = T) / mean(degree * weight2, na.rm = T),
            extortion = mean(degree[extortion == 1] * weight2[extortion == 1], na.rm = T) / mean(degree * weight2, na.rm = T),
            advanced_fee = mean(degree[advanced_fee == 1] * weight2[advanced_fee == 1], na.rm = T) / mean(degree * weight2, na.rm = T)) %>% 
  mutate(id = "id") %>% 
  pivot_longer(-id) %>% 
  select(scam = name, degree_ratio = value)

## Individual True Positive (Visibility Factors)
true_positive_oversample <- survey %>%  
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
  summarize(nonpayment = sum(visibility_nonpayment_count[nonpayment == 1] * weight2, na.rm = T) / sum(network_size[nonpayment == 1] * weight2, na.rm = T),
            advanced_fee = sum(visibility_advanced_fee_count[advanced_fee == 1] * weight2, na.rm = T) / sum(network_size[advanced_fee == 1] * weight2, na.rm = T),
            extortion = sum(visibility_extortion_count[extortion == 1] * weight2, na.rm = T) / sum(network_size[extortion == 1] * weight2, na.rm = T),
            overpayment = sum(visibility_overpayment_count[overpayment == 1] * weight2, na.rm = T) / sum(network_size[overpayment == 1] * weight2, na.rm = T)) %>% 
  mutate(id = "id") %>% 
  pivot_longer(-id) %>% 
  select(scam = name, true_positive_rate = value)

## store estimates for rare scams 
rare_adjustments <- degree_ratio_oversample %>% 
  left_join(true_positive_oversample) %>% 
mutate(prevalence = "rare") 

##########################################
#### Rare Pooled Adjustment Factors ######
##########################################

##  True Positive Rate
true_positive_oversample <- survey %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
  summarize(numerator_tp = sum(visibility_nonpayment_count[nonpayment == 1] * weight2[nonpayment == 1], na.rm = T) + sum(visibility_extortion_count[extortion == 1] * weight2[extortion == 1], na.rm = T) + 
              sum(visibility_advanced_fee_count[advanced_fee == 1] * weight2[advanced_fee == 1], na.rm = T) + sum(visibility_overpayment_count[overpayment == 1]* weight2[overpayment == 1], na.rm = T),
            denominator_tp = sum(network_size[nonpayment == 1] * weight2[nonpayment == 1], na.rm = T) + sum(network_size[advanced_fee == 1] * weight2[advanced_fee == 1], na.rm = T) + 
              sum(network_size[extortion == 1] * weight2[extortion == 1], na.rm = T) + sum(network_size[overpayment == 1] * weight2[overpayment == 1], na.rm = T)) %>% 
  mutate(true_positive = numerator_tp/denominator_tp)

## degree ratio 
degree_ratio_oversample <-  survey %>% 
  summarize(numerator_degree = mean(degree[nonpayment == 1 | overpayment == 1 | extortion == 1 | advanced_fee == 1] *
                                      weight2[nonpayment == 1 | overpayment == 1 | extortion == 1 | advanced_fee == 1], na.rm = T), 
            denominator_degree = mean(degree * weight2, na.rm = T)) %>% 
  mutate(degree_ratio = numerator_degree/denominator_degree)

## calculate adjustment factor 
rare_adjustment_factors <- bind_cols(true_positive_oversample, degree_ratio_oversample) %>% 
  mutate(adj_factor = 1/true_positive * 1/degree_ratio)    

##########################################

## survey general
survey_general <- survey %>% 
  filter(p_over == 1)

## top code 
survey_general <- survey_general %>%
  mutate(across(c(starts_with("name")), top_code))

###############################################
#### Common Individual Adjustment Factors  ####
###############################################

degree_ratio_genpop <- survey_general %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
                                       summarize(banking = mean(degree[banking == 1] * weight[banking == 1], na.rm = T) / mean(degree * weight, na.rm = T),
            banking_lostmoney = mean(degree[banking_lostmoney == 1] * weight[banking_lostmoney == 1], na.rm = T) / mean(degree * weight, na.rm = T),
            nondelivery = mean(degree[nondelivery == 1] * weight[nondelivery == 1], na.rm = T) / mean(degree * weight, na.rm = T)) %>% 
  mutate(id = "id") %>% 
  pivot_longer(-id) %>% 
  select(scam = name, degree_ratio = value)

true_positive_ratio <- survey_general %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9")) %>% 
  summarize(nondelivery = sum(visibility_nondelivery_count[nondelivery == 1]*weight[nondelivery == 1], na.rm = T) / sum(network_size[nondelivery == 1] * weight[nondelivery == 1], na.rm = T),
            banking = sum(visibility_banking_count[banking == 1]*weight[banking == 1], na.rm = T) / sum(network_size[banking == 1] * weight[banking == 1], na.rm = T),
            banking_lostmoney = mean(visibility_banking_count[banking_lostmoney] * weight[banking_lostmoney == 1], na.rm = T) / sum(network_size[banking_lostmoney == 1] * weight[banking_lostmoney == 1], na.rm = T)) %>%
  mutate(id = "id") %>% 
  pivot_longer(-id) %>% 
  select(scam = name, true_positive_rate = value)

## estimates for rare scams 
common_adjustments <- degree_ratio_genpop %>% 
  left_join(true_positive_ratio %>% 
              mutate(true_positive_rate = replace_na(true_positive_rate, 0)) ) %>% 
mutate(prevalence = "common") 

###############################################
#### Common Pooled Adjustment Factors      ####
###############################################

true_positive_genpop <- survey_general %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
  summarize(numerator_tp = sum(visibility_nondelivery_count[nondelivery == 1]*weight[nondelivery == 1], na.rm = T) + sum(visibility_banking_count[banking == 1] * weight[banking == 1], na.rm = T),
            denominator_tp = sum(network_size[nondelivery == 1] * weight[nondelivery == 1], na.rm = T) + sum(network_size[banking == 1] * weight[banking == 1], na.rm = T)) %>% 
  mutate(true_positive = numerator_tp/denominator_tp)

degree_ratio_genpop <-  survey_general %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
  summarize(numerator_degree = mean(degree[nondelivery == 1 | banking == 1  ] *
                                      weight[nondelivery == 1 | banking == 1], na.rm = T), 
            denominator_degree = mean(degree * weight, na.rm = T)) %>% 
  mutate(degree_ratio = numerator_degree/denominator_degree)

# combine adjustment factors ----------------------------------------------
common_adjustment_factors <- bind_cols(degree_ratio_genpop, true_positive_genpop) %>% 
  mutate(adj_factor = 1/true_positive * 1/degree_ratio)
```

## Bootstrap Uncertainty 
```{r}
# bootstrap estimates  -----------------------------------------------------

bootstrap_results <- list()

for (i in 1:100){
  
  ## survey_recode
  bootstrap_resample <- survey_subsample %>% 
    sample_n(nrow(survey_subsample), replace = T)
  

  ## manually compute scale-up estimates 
  resample_results <- bootstrap_resample %>% 
    summarize(banking = sum(scaleup_banking*weight)/sum(degree*weight),
              banking_lostmoney = sum(scaleup_banking_desc_lost_money_count*weight, na.rm = T)/sum(degree*weight),
              nondelivery = sum(scaleup_nondelivery*weight)/sum(degree*weight),
              nonpayment = sum(scaleup_nonpayment*weight)/sum(degree*weight),
              overpayment = sum(scaleup_overpayment*weight)/sum(degree*weight),
              advanced_fee = sum(scaleup_advanced_fee*weight)/sum(degree*weight),
              extortion = sum(scaleup_extortion*weight)/sum(degree*weight)) %>% 
    mutate(id = "id") %>% 
    pivot_longer(-id) %>% 
    select(scam = name, proportion = value)
  
  ## unweighted 
  degree_ratio_genpop <- bootstrap_resample %>% 
    summarize(banking = mean(degree[banking == 1] * weight[banking == 1], na.rm = T) / mean(degree * weight, na.rm = T),
              banking_lostmoney = mean(degree[banking_lostmoney == 1] * weight[banking_lostmoney == 1], na.rm = T) / mean(degree * weight, na.rm = T),
              nondelivery = mean(degree[nondelivery == 1] * weight[nondelivery == 1], na.rm = T) / mean(degree * weight, na.rm = T)) %>% 
    mutate(id = "id") %>% 
    pivot_longer(-id) %>% 
    select(scam = name, degree_ratio = value)
  
  true_positive_genpop <- bootstrap_resample %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
    summarize(numerator_tp = sum(visibility_nondelivery_count[nondelivery == 1]*weight[nondelivery == 1], na.rm = T) + sum(visibility_banking_count[banking == 1] * weight[banking == 1], na.rm = T),
              denominator_tp = sum(network_size[nondelivery == 1] * weight[nondelivery == 1], na.rm = T) + sum(network_size[banking == 1] * weight[banking == 1], na.rm = T)) %>% 
    mutate(true_positive = numerator_tp/denominator_tp)
  
  degree_ratio_genpop <-  survey %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
    summarize(numerator_degree = mean(degree[nondelivery == 1 | banking == 1  ] *
                                        weight[nondelivery == 1 | banking == 1], na.rm = T), 
              denominator_degree = mean(degree * weight, na.rm = T)) %>% 
    mutate(degree_ratio = numerator_degree/denominator_degree)
  
  # combine adjustment factors ----------------------------------------------
  common_adjustment_factors_resample <- bind_cols(degree_ratio_genpop, true_positive_genpop) %>% 
    mutate(adj_factor = 1/true_positive * 1/degree_ratio)
  
  ## true positive regular sample
  
  ## adjustment factor 
  resample_oversample <- survey %>% 
    sample_n(size = nrow(survey), replace = T)
  
  ## weighted True Positive Rate
  true_positive_oversample <- resample_oversample %>% 
  filter(!id %in% c("K9T5J5", "Y3A8A2", "O7M7F9", "K2Q0W1", "F6Y1X2")) %>% 
    summarize(numerator_tp = sum(visibility_nonpayment_count[nonpayment == 1] * weight2[nonpayment == 1], na.rm = T) + sum(visibility_extortion_count[extortion == 1] * weight2[extortion == 1], na.rm = T) + 
                sum(visibility_advanced_fee_count[advanced_fee == 1] * weight2[advanced_fee == 1], na.rm = T) + sum(visibility_overpayment_count[overpayment == 1]* weight2[overpayment == 1], na.rm = T),
              denominator_tp = sum(network_size[nonpayment == 1] * weight2[nonpayment == 1], na.rm = T) + sum(network_size[advanced_fee == 1] * weight2[advanced_fee == 1], na.rm = T) + 
                sum(network_size[extortion == 1] * weight2[extortion == 1], na.rm = T) + sum(network_size[overpayment == 1] * weight2[overpayment == 1], na.rm = T)) %>% 
    mutate(true_positive = numerator_tp/denominator_tp)
  
  degree_ratio_oversample <-  resample_oversample %>% 
    summarize(numerator_degree = mean(degree[nonpayment == 1 | overpayment == 1 | extortion == 1 | advanced_fee == 1] *
                                        weight2[nonpayment == 1 | overpayment == 1 | extortion == 1 | advanced_fee == 1], na.rm = T), 
              denominator_degree = mean(degree * weight2, na.rm = T)) %>% 
    mutate(degree_ratio = numerator_degree/denominator_degree)
  
  estimate <- bind_cols(true_positive_oversample, degree_ratio_oversample) %>% 
    mutate(adj_factor = 1/true_positive * 1/degree_ratio)
  
  bootstrap_results[[i]] <- resample_results %>% 
    mutate(estimate = case_when(
      scam == "banking" ~ proportion*common_adjustment_factors_resample$adj_factor,
      scam == "nondelivery" ~ proportion*common_adjustment_factors_resample$adj_factor,
      scam == "banking_lostmoney" ~ proportion*common_adjustment_factors_resample$adj_factor,
      TRUE ~  proportion* estimate$adj_factor)) %>% 
    mutate(resample_num = i)
}

bootstrap_results_df <- bind_rows(bootstrap_results) 

bootstrap_CI <- bootstrap_results_df %>% 
  group_by(scam) %>% 
  summarize(low = quantile(estimate, probs = 0.025, na.rm = T),
            upper = quantile(estimate, probs = 0.975, na.rm = T))
```


```{r save estimates}

gnsum_bootstrap_estimates <- basic_nsum_results %>% 
  mutate(proportion = case_when(
    scam == "banking" ~ proportion * common_adjustment_factors$adj_factor,
    scam == "banking_lostmoney" ~ proportion * common_adjustment_factors$adj_factor,
    scam == "nondelivery" ~ proportion * common_adjustment_factors$adj_factor,
    TRUE ~ proportion * rare_adjustment_factors$adj_factor
  )) %>% 
  inner_join(bootstrap_CI) %>% 
  mutate(scam = case_when(
    scam == "banking" ~ "Credit Card / Banking",
    scam == "nondelivery" ~ "Non-Delivery",
    scam == "banking_lostmoney" ~ "Credit Card / Banking (Lost Money)",
    scam == "advanced_fee" ~ "Advanced Fee",
    scam == "nonpayment" ~ "Nonpayment",
    scam == "extortion" ~ "Extortion",
    scam == "overpayment" ~ "Overpayment"
  )) %>% 
  mutate(method = "Generalized Scale-Up")

gnsum_bootstrap_estimates <- gnsum_bootstrap_estimates %>% 
  select(scam, method, estimate = proportion, low = low, upp = upper) %>% 
  mutate(estimate_annualized = 1 - (1-estimate)^(1/2),
         low_annualized = 1 - (1-low)^(1/2),
         upper_annualized = 1 - (1-upp)^(1/2))

gnsum_bootstrap_estimates <- gnsum_bootstrap_estimates %>% 
  mutate(scam = case_when(
    scam == "Credit Card / Banking" ~ "Banking / Credit Card",
    scam == "Credit Card / Banking (Lost Money)" ~ "Banking / Credit Card (Lost Money)",
    scam == "Nonpayment" ~ "Non-Payment",
    TRUE ~ scam
  ))

## scaleup adjustment factors pooled 
scaleup_adjustment_factors_individual <- common_adjustments %>% 
  bind_rows(rare_adjustments) %>% 
  mutate(`Adjustment Factor` = (1/degree_ratio) * (1/true_positive_rate)) %>% 
  mutate(scam = case_when(
    scam == "banking" ~ "Banking / Credit Card",
    scam == "nondelivery" ~ "Non-Delivery",
    scam == "banking_lostmoney" ~ "Banking / Credit Card (Lost Money)",
    scam == "advanced_fee" ~ "Advanced Fee",
    scam == "nonpayment" ~ "Non-Payment",
    scam == "extortion" ~ "Extortion",
    scam == "overpayment" ~ "Overpayment"
  )) %>% 
  select(Cybercrime = scam, Incidence = prevalence, `Degree Ratio` = degree_ratio, 
         Visibility_ratio = true_positive_rate, `Adjustment Factor`) 

scaleup_adjustment_factors_pooled <- common_adjustment_factors %>% 
  select(degree_ratio, true_positive_rate = true_positive) %>%  
  mutate(Incidence = "common") %>% 
  bind_rows(rare_adjustment_factors %>% 
              select(degree_ratio, true_positive_rate = true_positive) %>% 
              mutate(Incidence = "pooled")) %>% 
  mutate(adjustment_factor = (1/degree_ratio) * (1/true_positive_rate))

## write our adjustment factors 
write_csv(scaleup_adjustment_factors_pooled, "../data/results/nsum_pooled_adjustments_nofilter.csv")
write_csv(scaleup_adjustment_factors_individual, "../data/results/nsum_individual_adjustments_nofilter.csv")
write_csv(gnsum_bootstrap_estimates, "../data/results/gnsum_estimates_nofilter.csv")
```
